+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v0.5.0 (TBD)
 
+* Tracked token symbols with config file (#441).
 * [BREAKING] Refactored `TransactionRequest` to represent a generalized transaction (#438).
-* Fix `get_consumable_notes` to consider block header information for consumability (#432).
+* Fixed `get_consumable_notes` to consider block header information for consumability (#432).
 * Ignored stale updates received during sync process (#412).
 * Refactor `TransactionRequest` constructor (#434).
 * Fixed flaky integration tests (#410).